@@ -4,15 +4,15 @@ apply plugin: 'org.greenrobot.greendao'
 
 
 android {
-    compileSdkVersion 28
-    buildToolsVersion '28.0.3'
+    compileSdkVersion 29
+    buildToolsVersion '29.0.3'
     defaultConfig {
         applicationId "dev.ukanth.iconmanager"
         minSdkVersion 19
-        targetSdkVersion 28
-        versionCode 3110
+        targetSdkVersion 29
+        versionCode 200830
         versionName "3.1.1"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
         buildConfigField "boolean", "LICENSECHECK", "true"
     }
 
@@ -44,18 +44,21 @@ dependencies {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
 */
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    implementation 'com.android.support:cardview-v7:28.0.0'
-    implementation 'com.android.support:recyclerview-v7:28.0.0'
-    implementation 'com.android.support:support-v4:28.0.0'
-    implementation 'com.android.support:support-v13:28.0.0'
-    implementation 'com.android.support:design:28.0.0'
-    implementation 'com.android.support:support-annotations:28.0.0'
-    implementation 'com.android.support:support-vector-drawable:28.0.0'
-    implementation 'com.android.support:animated-vector-drawable:28.0.0'
-    implementation 'com.android.support:palette-v7:28.0.0'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'androidx.recyclerview:recyclerview:1.0.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.legacy:legacy-support-v13:1.0.0'
+    implementation 'com.google.android.material:material:1.0.0'
+    implementation 'androidx.annotation:annotation:1.0.0'
+    implementation 'androidx.vectordrawable:vectordrawable:1.0.0'
+    implementation 'androidx.vectordrawable:vectordrawable-animated:1.0.0'
+    implementation 'androidx.swiperefreshlayout:swiperefreshlayout:1.0.0'
+    implementation "androidx.constraintlayout:constraintlayout:2.0.0"
 
-    implementation 'com.android.support.constraint:constraint-layout:1.0.2'
+    implementation 'androidx.palette:palette:1.0.0'
+
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
 
     //root library
     implementation 'eu.chainfire:libsuperuser:1.0.0.201602271131'
