@@ -15,11 +15,10 @@ import android.graphics.drawable.BitmapDrawable;
 import android.graphics.drawable.Drawable;
 import android.os.AsyncTask;
 import android.os.Build;
-import android.support.annotation.NonNull;
-import android.support.v4.app.NotificationCompat;
+import androidx.annotation.NonNull;
+import androidx.core.app.NotificationCompat;
 import android.text.TextUtils;
 import android.util.Log;
-import android.view.View;
 import android.widget.Toast;
 
 import com.afollestad.materialdialogs.MaterialDialog;
@@ -42,7 +41,6 @@ import java.util.Random;
 import dev.ukanth.iconmgr.ApplyActionReceiver;
 import dev.ukanth.iconmgr.App;
 import dev.ukanth.iconmgr.DetailsActivity;
-import dev.ukanth.iconmgr.MainActivity;
 import dev.ukanth.iconmgr.Prefs;
 import dev.ukanth.iconmgr.PreviewActionReceiver;
 import dev.ukanth.iconmgr.R;
